requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
gspread>=5.10.0
google-auth>=2.22.0
